Massively optimise deconstruct() efficiency + progress bar
This commit introduces a shiny new progress bar for the deconstruct function, and I plan to use it in reconstruct() too. It took me too long to realise that I was loading files into memory every time, for small files this is acceptable, but testing with bigger files exposed the problem; nfl uses BufReader/BufWriter for deconstruct() now, and will soon use it for reconstruct() too.